[FM] MB-20158: Typecast size_t to 64-bit for printing on 32-bit
Add __printflike support to catch issues in fdb_log()
Remove unused _FSEC and _FUSEC macros
Unify _F64 to be the same on all platforms
Fix filemgr warning
filemgr.cc:486:59: warning: format string is not a string literal
(potentially insecure)

Reviewed-on: http://review.couchbase.org/65802
Tested-by: buildbot <build@couchbase.com>

+++ src/configuration.cc
@@ -144,7 +144,7 @@ bool validate_fdb_config(fdb_config *fconfig) {
         // Block size should be set between 1KB and 128KB
         fdb_log(NULL, FDB_RESULT_INVALID_ARGS,
                 "Config Error: Block size (%" _F64 ") not between 1KB and 128KB!\n",
-                fconfig->blocksize);
+                static_cast<uint64_t>(fconfig->blocksize));
         return false;
     }
